backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
psycopg2-binary
